+++ CHANGELOG.md
@@ -64,7 +64,7 @@
 - Remove interfaces `SaverInterface`, `BulkSaverInterface` and `RemoverInterface` from `Pim/Bundle/CatalogBundle/Manager/ProductManager` and remove methods `save`, `remove` and `saveAll`
 - Remove interfaces `SaverInterface` and `RemoverInterface` from `Pim/Bundle/CatalogBundle/Manager/AttributeOptionManager` and remove methods `save` and `remove`
 - Remove interfaces `SaverInterface` and `RemoverInterface` from `Pim/Bundle/CatalogBundle/Manager/CategoryManager` and remove methods `save` and `remove` and remove the $eventDispatcher argument
-- Remove interface `RemoverInterface` from `Pim/Bundle/CatalogBundle/Manager/GroupManager` and remove method `remove`
+- Remove interface `RemoverInterface` from `Pim/Bundle/CatalogBundle/Manager/GroupManager` and remove method `remove` and remove the $eventDispatcher argument
 - Remove interface `SaverInterface` from `Pim/Bundle/CatalogBundle/Manager/ChannelManager` and remove method `save`
 - Remove interfaces `SaverInterface` and `RemoverInterface` from `Pim/Bundle/CatalogBundle/Manager/FamilyManager` and remove methods `save` and `remove`
 - Change of constructor of `Pim/Bundle/CatalogBundle/Manager/FamilyManager` and remove `CompletenessManager`